+++ next_action/arguments/__init__.py
@@ -4,7 +4,7 @@ import argparse
 import os
 import shutil
 import sys
-from typing import Dict, List, Set
+from typing import cast, Dict, List, Set, Tuple
 
 from .arguments import Arguments
 from .parser import build_parser, parse_remaining_args
@@ -15,8 +15,8 @@ def parse_arguments() -> Arguments:
     # Ensure that the help info is printed using all columns available
     os.environ['COLUMNS'] = str(shutil.get_terminal_size().columns)
     default_filenames = ["todo.txt"]
-    arguments = Arguments(default_filenames)
     parser = build_parser(default_filenames)
+    arguments = Arguments(parser, default_filenames)
     namespace, remaining = parser.parse_known_args()
     parse_remaining_args(parser, remaining, namespace)
     arguments.filenames = namespace.filenames