+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The `--file` argument accepts `-` to read input from standard input. Closes #42.
+
+### Changed
+
+- Give consistent error message when files can't be opened. Fixes #54.
+
 ## [0.2.1] - 2018-05-16
 
 ### Changed